slider completly seperated from home component
