+++ src/components/Slider.js
@@ -1,10 +1,48 @@
-import React from 'react'
+import React,{useState} from 'react'
 import "./Slider.css"
+import CaseStudies from "../CaseStudies"
 
-function Slider1({ appInfo, image }) {
+
+let counter = 0
+let animate = "left-to-right"
+function Slider() {
+    const [sliderComponent, setSliderComponent] = useState({ comp: Slides[0] })
+    const nextSlide = () => {
+        counter++;
+        if (counter > CaseStudies.length - 1) {
+            counter = 0;
+        }
+        animate = "left-to-right"
+        setSliderComponent({ comp: Slides[counter % 2] })
+        
+    }
+    const prevSlide = () => {
+        counter--;
+        if (counter < 0) {
+            counter = CaseStudies.length - 1;
+        }
+        
+        animate = "right-to-left"
+        setSliderComponent({ comp: Slides[counter % 2] })
+    }
     return (
-        <div className='slider'>
-            <img src={image.address} className="slider-image" alt={image.alt}/>
+        <div className='case-studies-container'>
+            <i className='bi bi-chevron-left' onClick={nextSlide}></i>
+            <div className='slider-container'>
+                {<sliderComponent.comp
+                    image={CaseStudies[counter].image}
+                    appInfo={CaseStudies[counter].appInfo}
+                    animate = {animate}
+                />}
+            </div>
+            <i className='bi bi-chevron-right' onClick={prevSlide}></i>
+        </div>
+    )
+}
+function Container1({ appInfo, image ,animate}) {
+    return (
+        <div className={'slider '+animate}>
+            <img src={image.address} className="slider-image" alt={image.alt} />
             <div className='slider-text'>
                 <h3 className='slider-header-2'>{appInfo.header}</h3>
                 <h2 className='slider-header-1'>{appInfo.title}</h2>
@@ -15,10 +53,10 @@ function Slider1({ appInfo, image }) {
         </div>
     )
 }
-function Slider2({ appInfo, image }) {
+function Container2({ appInfo, image ,animate}) {
     return (
-        <div className='slider'>
-            <img src={image.address} className="slider-image" alt={image.alt}/>
+        <div className={'slider '+animate}>
+            <img src={image.address} className="slider-image" alt={image.alt} />
             <div className='slider-text'>
                 <h3 className='slider-header-2'>{appInfo.header}</h3>
                 <h2 className='slider-header-1'>{appInfo.title}</h2>
@@ -29,6 +67,6 @@ function Slider2({ appInfo, image }) {
         </div>
     )
 }
-const Slider = [Slider1, Slider2];
+const Slides = [Container1, Container2];
 
 export default Slider
